fix Struct & createGraph

+++ world/src/main/scala/com/sberhack/fun/car/CarNextMove.scala
@@ -1,14 +1,12 @@
 package com.sberhack.fun.car
 
 case class CarNextMove(
-                        nodeId: Long,
+                        nodeId: Int,
                         cashIn: Boolean
                       ) {
-
   override def toString: String = if (cashIn) {
     s"Move to Node $nodeId for cash in"
   } else {
     s"Move to Node $nodeId "
   }
-
 }
